moving MLP with modified weight init to common_models
