+++ nupic/research/frameworks/pytorch/models/common_models.py
@@ -25,6 +25,15 @@ from nupic.research.frameworks.pytorch.models.le_sparse_net import (
     add_sparse_linear_layer,
 )
 
+_all__ = [
+    "StandardMLP",
+    "ModifiedInitStandardMLP",
+    "SparseMLP",
+    "OmniglotCNN",
+    "OMLNetwork",
+    "MetaContinualLearningMLP"
+]
+
 
 class StandardMLP(nn.Module):
 
@@ -48,6 +57,40 @@ class StandardMLP(nn.Module):
 
     def forward(self, x):
         return self.classifier(x)
+
+
+class ModifiedInitStandardMLP(StandardMLP):
+    """
+    A standard MLP which differs only in feed-forward weight initialization: the bounds
+    of the Uniform distribution used to initialization weights are
+
+    +/- 1/sqrt(I x W x F)
+
+    where I is the density of the input for a given layer, W is always 1.0 (since MLPs
+    have dense weights), and F is the number of inputs received. This only differs from
+    Kaiming initialization by incorporating input_density and weight_density. Biases
+    are unaffected.
+    """
+
+    def __init__(self, input_size, num_classes, hidden_sizes):
+        super().__init__(input_size, num_classes, hidden_sizes)
+
+        # Modified Kaiming weight initialization which considers 1) the density of
+        # the input vector and 2) the weight density in addition to the fan-in
+
+        weight_density = 1.0
+        input_flag = False
+        for layer in self.classifier:
+            if isinstance(layer, nn.Linear):
+
+                # Assume input is fully dense, but hidden layer activations are only
+                # 50% dense due to ReLU
+                input_density = 1.0 if not input_flag else 0.5
+                input_flag = True
+
+                _, fan_in = layer.weight.size()
+                bound = 1.0 / np.sqrt(input_density * weight_density * fan_in)
+                nn.init.uniform_(layer.weight, -bound, bound)
 
 
 class SparseMLP(nn.Sequential):
